[Feat/#88] 네이밍 변경 및 API 추가
